+++ Cargo.toml
@@ -187,7 +187,8 @@ getrandom_v02 = { package = "getrandom", version = "0.2", features = ["js"] }
 getrandom_v03 = { package = "getrandom", version = "0.3", features = ["wasm_js"] }
 # No default features on WASM: excludes sqlite_cookie_store (needs C SQLite) and parley_layout (needs freetype).
 gosub_engine = { path = "./crates/gosub_engine", default-features = false, features = ["backend_vello"] }
-gosub_render_pipeline = { path = "./crates/gosub_render_pipeline", features = ["backend_vello"] }
+gosub_render_pipeline = { path = "./crates/gosub_render_pipeline" }
+gosub_renderer_vello = { path = "./crates/gosub_renderer_vello" }
 gosub_shared = { version = "0.1.1", path = "./crates/gosub_shared", registry = "gosub" }
 parking_lot = { workspace = true }
