+++ bin/gosub-screenshot/Cargo.toml
@@ -9,7 +9,8 @@ path = "main.rs"
 
 [dependencies]
 gosub_engine = { path = "../../crates/gosub_engine", features = ["backend_vello"] }
-gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline", features = ["backend_vello"] }
+gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline" }
+gosub_renderer_vello = { path = "../../crates/gosub_renderer_vello" }
 vello = { workspace = true }
 image = { workspace = true }
 once_cell = { workspace = true }
