+++ crates/gosub_render_pipeline/Cargo.toml
@@ -29,13 +29,6 @@ gosub_interface = { version = "0.1.2", path = "../gosub_interface", registry = "
 gosub_fontmanager = { version = "0.1.0", path = "../gosub_fontmanager", registry = "gosub" }
 gosub_lattice = { version = "0.1.0", path = "../gosub_lattice", registry = "gosub" }
 
-# backend_vello optional deps
-vello = { workspace = true, optional = true }
-wgpu = { workspace = true, optional = true }
-fontique = { version = "0.10.0", optional = true }
-hashbrown = { version = "0.17.1", optional = true }
-skrifa = { version = "0.42.1", optional = true }
-
 [target.'cfg(target_os = "linux")'.dependencies]
 gdk4-wayland = { workspace = true, features = ["wayland_crate"], optional = true }
 gdk4-x11 = { workspace = true, optional = true }
@@ -46,7 +39,6 @@ gosub_css3 = { path = "../gosub_css3", registry = "gosub" }
 url = { workspace = true }
 
 [features]
-backend_vello = ["dep:vello", "dep:wgpu", "dep:fontique", "dep:hashbrown", "dep:skrifa"]
 parley_layout = []
 wayland = ["gdk4-wayland"]
 x11 = ["gdk4-x11"]
