+++ crates/gosub_render_pipeline/src/render/backends.rs
@@ -1,6 +1,6 @@
 /// Default backend that doesn't render or return anything.
+///
+/// Concrete rendering backends (Cairo, Skia, Vello) live in their own
+/// `gosub_renderer_*` crates; the pipeline only defines the `RenderBackend`
+/// trait and ships this null implementation.
 pub mod null;
-
-/// Backend using the Vello graphics library.
-#[cfg(feature = "backend_vello")]
-pub mod vello;
