+++ crates/gosub_renderer_vello/src/lib.rs
@@ -1,6 +1,8 @@
+pub mod backend;
 pub mod compositor;
 pub(crate) mod font;
 pub mod rasterizer;
 
+pub use backend::{VelloBackend, WgpuContextProvider, WgpuResources};
 pub use compositor::{VelloCompositor, VelloCompositorConfig};
 pub use rasterizer::VelloRasterizer;
