+++ crates/gosub_renderer_vello/src/rasterizer.rs
@@ -5,8 +5,9 @@ use gosub_render_pipeline::common::texture::TextureId;
 use gosub_render_pipeline::common::TextureStore;
 use gosub_render_pipeline::painter::commands::PaintCommand;
 use gosub_render_pipeline::rasterizer::Rasterable;
-use gosub_render_pipeline::render::backends::vello::WgpuResources;
 use gosub_render_pipeline::tiler::{Tile, TileId};
+
+use crate::backend::WgpuResources;
 use parking_lot::Mutex;
 use std::sync::Arc;
 use vello::kurbo::{Affine, Rect, Vec2};
