+++ examples/egui-vello/Cargo.toml
@@ -9,7 +9,8 @@ path = "main.rs"
 
 [dependencies]
 gosub_engine = { path = "../../crates/gosub_engine", features = ["ui_eframe", "backend_vello"] }
-gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline", features = ["backend_vello"] }
+gosub_render_pipeline = { path = "../../crates/gosub_render_pipeline" }
+gosub_renderer_vello = { path = "../../crates/gosub_renderer_vello" }
 eframe = { workspace = true, features = ["wgpu"] }
 egui = { workspace = true }
 vello = { workspace = true }
