+++ src/wasm/renderer.rs
@@ -6,7 +6,7 @@ use gosub_engine::tab::TabDefaults;
 use gosub_engine::zone::{ZoneConfig, ZoneServices};
 use gosub_engine::GosubEngine;
 use gosub_render_pipeline::render::backend::ExternalHandle;
-use gosub_render_pipeline::render::backends::vello::VelloBackend;
+use gosub_renderer_vello::VelloBackend;
 use gosub_render_pipeline::render::{DefaultCompositor, Viewport};
 use gosub_shared::tab_id::TabId;
 use parking_lot::RwLock;
